requirements.txt
requests==2.31.0
asyncio==3.4.3
Pillow==10.0.0
pytest==7.4.0
pytest-benchmark==4.0.0